Components: Decouple and refactor how visual states are handled
This commit decouples state usage from 1 state per prop, to using a generic 'state' prop that accepts a variety of states. This translates to cleaner code for JS and CSS.

The only difference is helpText has to be entered in via the `helpText` prop vs. inline with the previous state prop. That's okay, as it offers more flexibility.

This pattern has been updated for `Input`, `Select`, and `Text` components.

Storybook + Tests + Docs have been updated to match.