Add HS-App theme stylesheet for components
This update adds an `hs-app` stylesheet (.scss/.css) for Blue. The idea for this is to allow for Blue components to be seamlessly dropped into `hs-app` with no/minimal issues in UI consistency.

Typically, it should be up to the consumer of this package (`hs-app`) to write the custom CSS necessary to override Blue, as Blue should be application agnostic. However, I think this approach is fine, and is the easiest/fastest way for us to start integrating into multiple projects.

Minimal adjustments were made to the actual React JS components. Most of these changes are SCSS based, and are isolated within the new `hs-app` theme directory under `src/styles/themes/hs-app`.

The idea for this came about during a call I had with @knicklabs.