Choice: Fix ID generation
This update fixes the `uniqueID()` generation for the `Choice` component, which indirectly fixes it for `Checkbox` and `Radio`. The same fix was also applied to `Input` and `PortalWrapper`.

This issue was the element's ID was being regenerated everytime there was a prop change, which is incorrect. It should only be generated once. The solution was to generate and store the uniqueID in the component state, instead of creating it in the render hook.

The uniqueID creation factory function has also been enhanced to accept a prefix override namespace.