CSS: Create hs-app dedicated stylesheet without duplicate styles
This update adds a new `blue.hs-app.slim.scss` file that only includes not within `hs-app`. This will reduce the overall build size of `hs-app`'s CSS.

The approach I took was to comment out all of the component imports in the dedicated slim file. It's not ideal, as you'd have to update 2 .scss files everytime a new component style is added. However, it's the simplest solution for now.

To integration into hs-app, simply import the slim file:

```
@import "blue.hs-app.slim";
```

Whereas prior to this, you needed import the base Blue styles + hs-app theme.