Implement enum style propTypes
This update provides most components with a specific range of acceptable values.

For example, the states (`string`) that we use in Blue are `error`, `success`, `warning`. Components with a `state` prop should only accept these values, as other values will fall to render the correct styling.

To be things easier (and more consistent), I've added `constants/propTypes.js` that contain an export of common propTypes, such as state or size.