Add SidebarCollapsibleCard component
**Add SidebarCollapsibleCard + Collapsible component*

This update adds a new general `Collapsible` component. It is self contained, and responsible for rendering content with the ability to expand/collapse. It also provides `onOpen` and `onClose` callback methods that can be tapped into.

This update also adds a new `SidebarCollapsibleCard` component, which uses `Collapsible` to handle the main functionality. I was able to re-use a lot of other Blue components (like `Heading` and `Flexy`) in constructing this sucker, so that's awesome!

**Add new caret icons**

This update also adds 4 new caret icons: `up`, `down`, `left`, `right`.

Tests and readmes were added/updated for all the things!

This update mostly adds additional functionality and does not modify previous component API. Users/consumers of Blue should be unaffected.