Add EmojiPicker and Emoji components
This update adds 2 new components: `<EmojiPicker />` and `<Emoji />`. Both of these components are light wrappers around the components provided by the [emoji-mart](https://github.com/missive/emoji-mart/) library.

The strategy I used to tweak the styles EmojiPicker was to structure the `EmojiPicker/*.scss` files **as if we had smaller components**, exactly like how we structure the CSS for other Blue components.

Each `.scss` file under `EmojiPicker` is scoped under the main component class. This allows the styles to not leak into global scope, and it allows for us to easily override the default library styles in an easy-to-understand + maintainable way.