Add InfiniteScroller component + enhance Scrollable
This update adds a new InfiniteScroller component that makes it easy to
dynamically fetch/load/inject content (or do anything else!). It has a couple
of handy callbacks to provide the user with control when the component is
scrolled into position.

The Scrollable component has also been enhanced with a new `onScroll` callback.