SidebarCollapsibleCard: Improve Sortable interactions
This update fixes Sortable's handling of child item keys. Sortable now respects unique keys, and uses ID props if specified.

SidebarCollapsibleCard gains 2 new hooks: onSortStart and onSortEnd that fire when the component is being sorted. The react-sortable-hoc did NOT provide any callback methods for these hooks for the DraggableHandle, only at the Container level. So I had to come up with a creative solution to get this to work.

The two props allow for SidebarCollapsibleCard to be better integrated into Sortable.

Collapsible gains 2 new props: durationOpen and durationClose, that fallback to duration. This allows you to pass in different timings for the animation, allow for a smoother collapse experience with SidebarCollapsibleCard.