KeypressListener: Work with iframes
This update adds a new node util, that retrieves the closest `document`, given
a DOM node. This allows for KeypressListener to be used within an `iframe`,
and bind events to the iframe `document`, vs the global window `document`.

The scope is automatically set in the KeypressListener component. Various
other Blue components that use these bindings like Portal, Modal, or Dropdown,
do not need to be updated.