Add Switch and Hr components
This update adds two new components to Blue: Switch and Hr.

The Switch component is an alternative component to Checkbox, and provides
a switch-toggle UI.

The Hr component is an enhanced verson of the HTML default `hr` selector.

The styling for both components are provided by Seed (`seed-switch` and
`seed-hr`). These packages were installed as devDependencies, which
cause and update to `package-lock.json`.

Stories, Tests, and README have been added for both components.

These components are presentional, and don't affect any other components.