Input: Fix webkit autocomplete styles
This update adjusts the CSS for Input (and indirectly, Select), to allow for
the webkit autocomplete state to render correctly.

Previously, the autocomplete styles were overlapping the Input component's
"Backdrop" layer (a technique used to render the appearance/border of the
component). To resolve this, I've added a 1px padding around the container,
which simulates the offset a typical non-backdrop border would provide.

With this change, the autocomplete styles no-longer overflow on top of
the backdrop layer.

This update is purely a CSS change. The only JS changes made were to update
the Input component Stories to add an autocomplete example (and to refactor).

:clapping: