Dropdown: Simplify positioning
This update simplifies the positioning calculations used by the Dropdown
component (more accurately, the Drop.Positioner component).

The adjusted positions are similar to the ones available in other
Dropdown libraries, like Bootstrap.

This is a temporary "downgrade", until I can implement the 12/16-point
positioning system, that we can leverage for Tooltips and Popovers.

For now, the Drop.Positioner calculations only benefit Dropdown.

Note: Up/down recalcuation is still available. However, left/right
recalculations have been removed.

**Bug fixes**

The DropdownMenu has also been enhanced to only `scrollIntoView` if the
Scrollable component has a scrollbar. This is to resolve that function
forcing the window to scroll.