Add Timeline component + Refactor Timestamp
This commit adds a new Timeline component, which renders list items
within a timeline UI.

This commit also abstracts Timestamp to it's own component, since it's
being shared in both Message and Timeline. In order for the hover
interaction to be shared, a new HoverWrapper HOC was created, which
wraps both Message.ChatBlock and Timeline.Item.

A new DetailList component was also added, that serves as a light-weight
wrapper for the default HTML `dl`.

Some tests for Messages was refactored for the Timestamp abstraction
and for the new HoverWrapper HOC.