Adds AvatarGrid and StatusDot components
This update adds a new AvatarGrid component. It's similar to
AvatarStack and AvatarList, in that it renders an array of Avatars.
However the presentation for the avatars is different. Like AvatarList,
the new AvatarGrid component provides animations.

**Example**

```jsx
<AvatarGrid>
  <Avatar />
  <Avatar />
  <Avatar />
</AvatarGrid>
```

Eventually, we'll want to update the AvatarStack and AvatarList
components to follow the same `children` prop convention, instead
of passing in Avatars via an `avatar` prop.

This update also adds a new StatusDot component, which is now
built into Avatar, and can be enabled by passing in the appropriate
`status` prop.

**Example**

```jsx
<Avatar status='offline' />
```