Pressing ESC should not close all Portal components
This update adds a new "Manager" to the PortalWrapper component,
created with the new `globalManager` utility. This manager ensures
that pressing Escape ONLY closes the most recently Portal'ed component.

This is done by the manager tracking the IDs (which are unique) of
all the PortalWrapper created components.

This has been tested with Modals as well as Dropdowns.