StatusDot: Add additional props
This update enhances the StatusDot component by providing it with
additional props to have a more varied rendering style.

Additional props include:

* `inline` (ability to render the dot as `display: inline-block`)
* `size` (`md`|`sm`)
* `outerBorderColor`
* a new `active` state (which renders a green dot)

The new `active` state allows for this component to be more general,
instead of using `online` as the state for a green color.

Documentation and tests have been updated.