StatusDot: Icon
This update adds a new Icon prop to the StatusDot component, allowing it
to render an Icon component within, but also to adjust the component's
style.

This will add additional flexibility for the use-cases for the StatusDot
component :).

Note: You can pass any icon you wish into StatusDot (assuming it exists
as an Icon component `name`). The color of the icon is based on the
`status` of the StatusDot.

For example, if you want a red `x` icon, you would write:

```jsx
<StatusDot status='offline' icon='cross' />
```

To make it easier to use some of the more common icons, icons like
`tick-large` or `cross-large` have received aliases of `tick`, `cross`,
etc...

No updates are required for the user's end.

Tests + readme have been updated.