Modal: Fix overlapping UI for scrollable content
This update adjusts the width of the Scrollable faders and the position
of the CloseButton for Modal. Previously, these elements had fixed
sizes/offsets, which overlapped with the (native) scrollbar that appears.

To remedy this, a util function was added to calculate the web
scrollbar width, determined by the OS. Calculcations where then added
to inline styles when applicable.

This update also fixes the Modal scroll bug, introduced by the
AnimeJS integration.