StatusBar: Add themes, status, and Button
This update specifies and adds two themes to StatusBar: `light` and
`bold`. The `light` theme will be the default theme used for
the component. The color variations have also been enhanced by
adding an array of `status` choices (`info`, `success`, `error`,
and `warning`).

The component is also enhanced with the ability to not close on
click via the `closeOnClick` prop.

Lastly, a new sub-component have `<StatusBar.Button />` has been
added provide a button UI within the StatusBar.