Tag: Enable ability to be removed
This update enhances Tags with the ability to be removed from the DOM
by clicking an `x` Icon.

To properly contain and manage a list of tags, two new components were
added: `TagList` (used to contain Tags), and `Inline` (used for
controlling layout and overflow).

Similar to the other list-based components, Tags can be added
into a TagList like so:

```jsx
<TagList>
  <Tag color='red'>stay</Badge>
  <Tag color='green'>class</Badge>
  <Tag color='grey' filled>san diego</Badge>
</TagList>
```

An example of data flow for removing tags has been added to the
TagList storybook.

Toolbar also receives a more realistic example of a TagList working
next to an AvatarList.