PortalWrapper: Add wrapperClassName
This update adds a new `wrapperClassName` prop for the composed
components generated by PortalWrapper. This allows you to add
custom class names to the surrounding Portal'ed `div`, providing
extra CSS flexibility (if needed).

This is an enhancement, and no component API changes are required
on the user's end.

Tests + documentation has been added for this update :rocket: