Modal: Add Header/Body/Footer sub-components
This update adds a new sub-components to Modal: Header, Body, and Footer.
This Header and Footer sub-components are constructed with Toolbar,
and allow you to easily add items (typically action items) to a
Modal dialog.

**Example**
```jsx
<Modal>
  <Modal.Header>
    ...
  </Modal.Header>
  <Modal.Body>
    ...
  </Modal.Body>
  <Modal.Footer>
    ...
  </Modal.Footer>
</Modal>
```

The Modal component has been adjusted to accommodate the new
required markup structure. It is recommended that content is added
into Modal using a `Modal.Body` wrapper. However, if one is not
provided, the Modal component will automatically wrap the content for
you. This is so that the responsive behaviour of the Modal
(powered by CSS flexbox) is preserved.

The docs have been updated to encourage use of Modal.Body.