Modal: Add Content sub component
This update adds a new Modal.Content sub-component for Modal.
The idea for the sub-component is to support swapping Modal's `children`
from one set of components to another (e.g. loading conditions).

As if this update, Modal will no longer render anything inside that
isn't a sub-component. This is to enforce the new component
pattern required.

This update makes it slightly more verbose to add Modal content.
However, it enables better and more reliable flexibility in rendering
various types of content.

**Note to consumers of Blue: This update changes the Modal component
API. Please refer to the docs (or see examples below) if you notice
that your Modals are not rendering content.**

**Example**
```jsx
<Modal>
  <Modal.Content>
    ...
  </Modal.Content>
</Modal>
```

This will still work:
```jsx
<Modal>
  <Modal.Header>
    ...
  </Modal.Header>
  <Modal.Body>
    ...
  </Modal.Body>
</Modal>
```

This will no longer work:
```jsx
<Modal>
  Content
</Modal>
```