Modal and List: Bug Fixes
This update is a follow-up to the Modal.Content sub-component addition.
When implementing the new Modal structure, I discovered by rejecting
non-sub-component children, it causes unexpected issues for wrapper-like
components.

**This restriction has been removed for Modal.**

It is still highly recommended that you use either Modal.Body or Modal.Content to ensure that the UI looks correct.

There was a CSS bug that was introduced into Lists during the Tag
enhancement, which changed the display from `block` to `flex`.

This update fixes this by defaulting to a display of `block`, but
providing the ability to use `flex` via a prop.