CloseButton enhancements and Overflow fixes
This update adds hover, active, and focus CSS styles to the CloseButton.
A storybook example for CloseButton has also been setup to demo these
new styles, as well as the various sizes CloseButton provides.

Also in this update are fixes for Overflow, which was affected when used
with Tags. Overflow has been updated to better handle child components
wrapped with Animate. An additional check is done on mount via a `setTimeout`,
which can be adjusted, to check and re-adjust the height of Overflow.
This is useful for things wrapped in Animate, as they typically don't exist
in the DOM for the first couple of render cycles.

Lastly, some enhancements were made to the Tag and Badge components,
specifically adding a `display` prop which enables them to render with
`display: block` or `display: inline-block`. For certain UI, this was
useful to ensure consistent UI height. This was spotted when testing out
the ChatList.Item component.