Avatar: Adjust sizing
This update adjusts the `md` sizing for the Avatars to match the
pattern library. The Avatar components have also been provided with a default
`size` prop of `md`.

Lastly, `AvatarStack` has been enhanced with a new prop, `avatarClassName`.
This allows you to pass in a custom className to the child Avatars.