Illos: Add new illos and enhance :sparkle:
This update adds a handful of new Illo SVGs for the `<Illo>` component.
It also provides a new size setting of `72`.

This most exciting update is to provide illo with the ability to customize
desired illo color (for applicable paths) :heart:. This is done with
the unique way the SVG paths have been setup to coordinate with the
component's CSS and React props (but mostly CSS).