Skeleton: Add Control
This update adds a new Skeleton component - `Skeleton.Control`.
This component acts as a placeholder for things like
Dropdowns, Selects, Inputs, and Buttons. It has been
set up with the `size` prop that all of the components
share for their control elements.

This update also fixes the red border issue for `InputField`
elements on Firefox.