Animations: CSS over JS
This update enhances the Animate component to use CSS (className) based
animations instead of relying on inline styles. This has proved to be
more reliable and reduces the styles clashing/overriding each other
as they try to consolidate.

This update also resolves some issues with the PortalWrapper HOC not
accepting the `timeout` prop.