SCSS: Build raw files in `dist` and adjust CloseButton
This update enhances the build process of Blue to export all of the raw
SCSS files, allowing the user to customize what styles to use in their
project, if they so choose.

This is aided by an `includePaths.js` file that the user can use for their
`node-sass` setup to resolve external path depenedencies.

The CloseButton's UI has also been adjusted to ensure proper alignment
and better colors.