Select: Fix placeholder for Firefox
This update fixes the Select component to properly render a placeholder
for Firefox. The fix was to include the placeholder text as textNode
for the `<option>` element.

Previously, it was only used as a `label` prop. Chrome/Safari recognized
this, but Firefox did not.

With this update, all 3 browsers recognize the placeholder just fine.