Input + Select: (Force) autoFocus
This update enhances the Input and Select components to be able to
(better) autoFocus when they mount. It also allows for the user to
set the focus state using the new `isFocused` prop.

Lastly, there's a `forceAutoFocusTimeout` prop that allows the user to
fine tune when they would like the `setTimeout` based autofocus function
to fire (which is required for this technique to work).