Color: Adjust values + provide ability to override
This update adjusts both the `getColorShade` and `optimalTextColor`
util functions to use a different RGB weight to (subjectively) better
results for determining whether the text color should be black or white
(given a color).

The functions have also been enhanced to accept a secondary argument
(object), that allows the user to override the default suggested RGB
values with their own (to fine tune).