Notification/NotificationStack Components
This update adds 2 new components, `Notification` and `NotificationStack`.
These are used to display notification-based UI, which can be used in
interactions like indicating to the user of new Chat messages.

For the most part, the UI of the Notification uses the existing `Message.Chat`
component. Modifications were done to make that component more flexible to
provide better re-usability.