Notification: Enhancements
This update enhances the `Notification` component by providing it with the
ability to align content to the left or to the right.

The CSS styles that rendered the Notification UI has been abstracted from
`Message/Bubble.scss` to `Notification.scss`.

Lastly, truncation has been added for Notification body text. This is
to provide giant bubbles from appearing if there is too much content.
For added flexibility, the truncation limit can be customized via the
new `truncateLimit` prop.