Message: Add states and Spinner
This update adds error and loading states to the Attachment, Media,
and Chat sub-components of `Message`. The UI is handled by the `Chat`, which
exposes props for the wrapping `Attachment` and `Media` components to
retrofit their desired error/uploading UI.

A new Spinner component has been added to Blue, to help support the
"Uploading" state of the Message components. The Spinner component uses
the new `spinner` SVG Icon from `<Icon>`, and spins it with CSS animations.

Lastly, this update fixes a CSS regression that caused `from` Messages to
not have the correct border-radius styles.

Tests and README docs have been updated.