Attachment: Download should open in new window
This update enhances the `<Attachment>` component to automatically fill
the `download` and `target` HTML link attributes if a `url` is provided.
This allows for the attachment to reliably download/open in a new tab.

Download handling is custom for every browser. For Chrome's case, it
favours simply loading the file in a new tab.

To customize the behaviour, you can provide your own `download`,
`target` attributes, which will override the auto-generated ones.

Example:

**Usage**

```jsx
<Attachment link='https://github.com/helpscout/blue/raw/master/images/Blue.png' />
```

**Result**

```html
<a class="c-Attachment is-link" href="https://github.com/helpscout/blue/raw/master/images/Blue.png" download="" target="_blank">
  ...
</a>
```