ChatSidebar: Visual tweaks
This update adjusts the visual UI of various components that appear
within the ChatSidebar UI, which includes:

* Avatar
* Badge
* ChatInbox
* ChatInbox/Header
* ChatList
* ChatList/BlankSlate
* ChatList/Item
* ChatList/Header
* Tag
* Toolbar

Most of these changes are CSS changes. However, there were some minor
JS changes.

The most significant JS changes include the addition of a `xs` prop
for `Avatar` components, which should be used within the context of a
ChatSidebar.

Also, the example story for `ChatList` has been resolved to properly
render the Assigned UI.