StatusAvatar: Add Component + Enhancements
This update adds a new StatusAvatar component, which is an enhanced version
of `Avatar`, making it easier to indicate a user's state when presenting
an Avatar.

This update also adds 2 new icons to the `<Icon>` component,
`arrow-down-huge` and `arrow-up-huge`.

Lastly, there are tiny adjustments/propType fixes for components like `Tag`,
`AvatarList`, and `Avatar`.