Fancy 2.0 Update
This update bumps Fancy to the latest v2.0.0, which now uses an enhanced
fork of Emotion under the hood.

This allows for a more robust theming system, as well as iFrame support for
components that appear inside/outside the iFrames.

Theming/Scoping/iFrames (and any of those combos) have been thoroughly
tested within Fancy as well as Fancy's integration with Blue for all
major browsers, including IE 11.