Message.Media/Image: Responsive aspect ratio support
This update enhances the `Image` component to automatically calculate
aspect ratio sizing if `maxHeight` and `maxWidth` are defined.

This allows for the `img` to have a defined dimension before it renders into
the DOM. This interaction benefits UI like `Message.Media`, as content will
no longer jump around while the image attempts to render.