Message.Content: Note theme
This update enhances the `Message.Content` component to add `isNote` theme
support.

This is done by leveraging the (polyfilled) React 16 `context` API, which sets
up a Provider/Consumer relationship between the `Message` component and the
`Card` component that renders within.

The activation of the note theme happens internally.

CSS has been refactored for the `PreviewCard` and `Message.Content` to use
Fancy instead of SCSS.

Lastly, the main `Message` component was converted from a SFC to a React.Class
component for a performance/stability boost.