Text/Heading: Add dynamic sizing
This update adds dynamic sizing for `Text`/`Heading` components, which
leverages CSS variables for compatible browsers with fallbacks for
non-supported browsers (IE).

The `Tooltip` component was also improved to better handle scenarios where
a `title` prop may not be present.

Lastly, there was a CSS fix implementation of linear-gradient for SkeletonBlock