AvatarGrid: Remove animation
This update fixes the (re)rendering issues for AvatarGrid, caused by
`AnimateGroup`. The `AnimateGroup` component works correctly. However, the
combination of it within the `AvatarGrid` UI may not be ideal.

The CSS styles for `AvatarGrid` now live in Fancy CSS-in-JS, and the original
SCSS removed.

Lastly, tests and the README doc were updated to reflect the changes.