Input/Select: Add error icon support
This update enhances the `Input` and `Select` components with error icon
rendering. If these components have a `state` (prop) of `error`, they will
automatically add an error `Icon` at the end.

These error `Icon` have the ability to render a `Tooltip` message, if
the prop `errorMessage` is provided.

The `Select` component was refactored with Flow (yay).

For this update, I held off refactoring the CSS of `Input` and `Select`
with Fancy (for now).

Other updates include swapping `shallow` for `mount` rendering in tests,
updating tests, and improving `Truncate`/`Tooltip`/`Pop` rendering.