Portal: Fallback render to CLOSEST document
This update fixes `Portal` to ensure that it renders to the CLOSEST
`document` relative to the mounted component, rather than `window.document`.
This ensures proper iFrame support, out of the box.

This is achieved by using `getDocumentFromComponent`, which is provided
by our very own `react-utils` library.

`react-utils` abstracts several utils originally written in Blue (with
some new ones). Because of this, I was able to remove several utils and
a bunch of test code from Blue (yay).