ConfigProvider: A configurable Blue. Unlocked!
This update adds a brand new component, the `ConfigProvider`. Powered by
React's new `context` API (polyfilled for React 15), it enables users to
specify configs to customize Blue components within an applicable.

The first component to use this is `Tooltip`, starting with just the `zIndex`
prop. I've wanted to implement this system for a while. The `zIndex`
update was the perfect excuse, as it is necessary for `Tooltip` components
to render correctly within nested `Modal`.

The `zIndex` can be customized at the App, or even component scope level
as required.