Input: Add onResize callback
This update adds `onResize` to the `Input` component, which fires whenever
the `multiline` textarea resizes.

`PropProvider` was also refactored and improved. It follows HOC conventions
established by libraries like `redux`, such as component wrapping and
hoisting statics.

`configConnect` was also renamed to `propConnect`, to better reflect
`PropProvider`. The version of `react-utils` was bumped,
and used dependencies from that library instead (such as `createContext` and
`hoistNonReactStatics`.