Avatar: Adjust StatusDot position + refactor
This update adjusts the StatusDot position when it is rendered within
an `Avatar`.

This update provided the opportunity to refactor `Avatar`, `AvatarGrid`,
`AvatarList`, and `AvatarStack` with the latest Blue conventions, such
as `propConnect`, namespacing, and `CSS-in-JS`.

The tests were updated to use `mount` based testing, and the SCSS files
have been cleared!