PropProvider: Enhance to extends prop configs
This update improves `PropProvider` to allow nested instances to **extend**
configs. This makes usage of PropProvider more versitile. Allowing it to be
used, not just at the root App level, but in views, sub-views, and component
levels as well.

This Provider component is constructed using patterns adapted from libraries
like Emotion and Styled Components. The have an internal Pub/Sub system that
allows them nested Providers to communicate to pass along configs internally.

The propConnect() component then subscribes to the closest PropProvider,
where the extended props are then merged and rendered.