Page: Refine and add responsive support
This update refines the visuals for the various Page components. Most
importantly, it adds responsive support in the form of a new `isResponsive`,
prop.

This prop can be passed individually to the sub-components for finer grain
control. Alternatively, it can be defined on `Page`, and it will automatically
be passed down using it's internal `PropProvider`.