Page + Control improvements (#366)
* Input/Select controls: Improve focus

This update improves the Focus styles for Input and Select. This was done
by updating and implementing the internal `BackdropV2` component.

This allows us to remove the the original Backdrop component, along with
it's unused SCSS code.

* Page: Refine and add responsive support

This update refines the visuals for the various Page components. Most
importantly, it adds responsive support in the form of a new `isResponsive`,
prop.

This prop can be passed individually to the sub-components for finer grain
control. Alternatively, it can be defined on `Page`, and it will automatically
be passed down using it's internal `PropProvider`.

* Update and fix tests for Page and Input