Skeleton: Refactor + CSS-in-JS (#387)
* Skeleton: Refactor + CSS-in-JS

This update refactors the various Skeleton components with the latest Blue
conventions, including namespacing, and propConnects, and renders the
styles with CSS-in-JS.

A new `withAnimations` prop has also been added. This allows you
to disable animations, if required.

* Add babel-plugin-emotion

* RadioCard: Fix checked UI

This update fixes a regression for RadioCard that was introduced with
the recent `Choice`/`ChoiceGroup` enhancements.