AvatarList: Fix duplicate avatars
This update fixes the duplicate Avatar issue for the AvatarList component.
The issue was due to how the component(s) handled keys.
The solution was to not use AnimateGroup (unfortunately), which allowed
for React to properly figure out unique keys during the max->clamp process
of the AvatarList.

AvatarList also received a nice refactor of it's methods as well as a
conversion from JS+Flow to Typescript.