Fancy: Update to use PureComponent by default (#415)
This update bumps Fancy to the latest version, which defaults to creating
PureComponent classes, which helps with performance.

Some minor adjustments were required for a couple of Message components,
mainly tests. This was good, as it revealed an implementation issue
with the `ThemeProvider `component.

A story was added for some manual testing of Message.context (just in case).