ChatScroller: Reset Nodes
This update allows for `ChatScroller` to reset it's target DOM node when
either the `scrollableNode` or `scrollableSelector` props are updated.

Previously, the method to set the DOM node only ran during initialization.
With this update, `ChatScroller` is now able to observe different nodes
during it's lifecycle, depending on the implementation.