SelectDropdown: Add support for value
This update improves the `SelectDropdown` to be able to accept a
`value` prop. This adds API parity with `Select` as well as the
native HTML select.

`value` will be used to find the actual item from the original `items` prop.

Both `selectedItem` and `value` is supported for `SelectDropdown`.

An enhancement was made to `Dropdown` to add a new `triggerStyle` prop,
allowing you to specify inline styles for the Trigger wrapper
component.