Tag/TagList: Add onBeforeRemove feature + refactor (#448)
This update adds a new onBeforeRemove callback, which is a Promise that can
be passed to resolve before the Tag is removed. This allows for the user
to coordinate the tag removal experience with async-based actions, like
making API fetches.

Tag, TagList, and Centralize have been refactored with the latest HSDS
conventions, which includes converting `.js` to TypeScript, as well as
the latest `propConnect`, `css.js`, and `classNames()` conventions.