AnimationGroup/CardList: Animation consolidation/rendering fixes
This update fixes the issue with `AnimateGroup`/`CardList` to properly
handle the programatic adding/removing of child components.

This was done by updating `react-transition-group` (which is React 15 safe)
to the latest update and leverage it's `childFactory` callback function.

The `CardList` story was enhanced to thoroughly test these scenarios.

Lastly, `CardList` was refactored to use Typescript :tada:.